requirements.txt
numpy
pandas
scipy
scikit-learn
matplotlib
networkx
rdflib
torch
pykeen>=1.10.2
umap-learn
hdbscan
python-igraph
leidenalg
xgboost
lightgbm
tqdm
pyyaml
openpyxl
shap
